Implemented jumps
Implemented jump jumpnz (jump if not zero). Allowing for conditional
execution of runk code. It is still pretty rough around the edges.
Especially with runk trying to bestfit expression results in combination
with the current typechecking system for functions.

- Added two new variants to var::Var these are T (text) and L (lable).
- Runk can now process lable literals and pass them as function arguments.
- Runk will remember where different lables were declared.

It is still not possilbe to create lable or text variables and text
literals are not implemented at all.